move data classes to model folder, add better tests
