+++ bardapi/models/user_content.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod, abstractproperty
+
+
+class UserContent(ABC):
+    @abstractmethod
+    def key(self) -> str:
+        pass
+
+    # markdown representation as a long string
+    @property
+    @abstractmethod
+    def markdown_text(self) -> str:
+        pass
