+++ tests/data.py
@@ -1,4 +1,6 @@
 # Show 5 pictures of orcas
+import json
+
 t1 = [[None, ['c_85b578db36624ef5', 'r_85b578db36624470'], [['5 pictures of orcas', 1], ['Why is orca called Killer?', 4], ['How many orca is left?', 4]], None, [['rc_85b578db36624acd', ["Here are 5 pictures of orcas:\n\n[Image of Orca jumping out of water]\n[Image of Orca swimming underwater]\n[Image of Orca pod swimming together]\n[Image of Orca calf swimming with its mother]\n[Image of Orca spyhopping]\n\nOrcas are highly intelligent and social animals. They live in groups called pods, which can range in size from a few individuals to over 100. Orcas are apex predators, meaning they are at the top of the food chain. They have a diverse diet, including fish, squid, seals, sea lions, and even other whales.\n\nOrcas are found in all of the world's oceans, from the Arctic to the Antarctic. They are particularly common in coastal waters, where they can find plenty of food. Orcas are also known to travel long distances, and some pods have been tracked migrating thousands of miles each year.\n\nOrcas are important members of the marine ecosystem. They help to keep populations of other marine animals in check, and they also play a role in nutrient cycling. Orcas are also a popular tourist attraction, and many people enjoy watching them in the wild."], [], None, [[[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca jumping out of water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca jumping out of water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca jumping out of water]', 0, 'Orca jumping out of water', None, None, 1]], [[['https://capturetheatlas.com/wp-content/uploads/2022/12/Orca-wave-bait-ball.jpg'], None, 900, 600, 'Orca swimming underwater'], [['https://capturetheatlas.com/swimming-with-orcas/'], 'capturetheatlas.com', 5, 'https://encrypted-tbn2.gstatic.com/favicon-tbn?q=tbn:ANd9GcSmt003l7J2VaqnZKIcSAyP5VyMvwr1_YZRA3kyH2Pl8X6uNC6lqvM-CduQSAJmaesYVLFQhX2cX3dnCHQM78ww8EqbNMImpwS-ReAqgw'], '[Image of Orca swimming underwater]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcThgtfk8FZsRWE_FqTBx1n85uzM0PFsALLrev_ypkpz36SApo9PZkl_fhgMWBFH'], None, 250, 200], 0, None, None, ['[Image of Orca swimming underwater]', 0, 'Orca swimming underwater', None, None, 1]], [[['https://www.mensjournal.com/.image/t_share/MTk4Njg1MDA0MDM5OTg4OTk5/killer-whales-monterey-bay.jpg'], None, 1200, 671, 'Orca pod swimming together'], [['https://www.mensjournal.com/news/pod-30-killer-whales-monterey-bay'], 'www.mensjournal.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcSBy9iRr5b6ly40d-TXvQ9mVhN0XAYWUbqXniDVvEGTXMdBUHpsZkD3bNsg9CQYvJzLsZjHYPLn7KuoFQE7C1b46bJ3eKURI4i6McGsIg'], '[Image of Orca pod swimming together]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcT_zbn3aSj3FAvlBoWAq3Fa-S1ZirOYsyG3daLL5b_VhN0_z7-fdLiTW9LwKQ6r'], None, 250, 200], 0, None, None, ['[Image of Orca pod swimming together]', 0, 'Orca pod swimming together', None, None, 1]], [[['https://i.dailymail.co.uk/1s/2021/05/13/12/42941438-0-image-a-13_1620906509879.jpg'], None, 636, 382, 'Orca calf swimming with its mother'], [['https://www.dailymail.co.uk/sciencetech/article-9574917/Orca-enthusiast-captures-incredible-photo-rare-newborn-calf-coast-Scotland.html'], 'www.dailymail.co.uk', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcTxhjGkCz1VjtErZLwi2Qoupz4JRVjuXvn5-XVV95YtBS9w8GoFJ36mkZqAG33esE0y2aqDVRmnxtwu7CbUzJFR5TGZCQDTQq-fCGJ-kA'], '[Image of Orca calf swimming with its mother]', [['https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSsAMSHd01qPovodn7BzzGKE_u1SpRwMLVrsWo_VexZRGf3PMipYd9gku8uhr78'], None, 250, 200], 0, None, None, ['[Image of Orca calf swimming with its mother]', 0, 'Orca calf swimming with its mother', None, None, 1]], [[['https://nplimages.infradoxxs.com/cache/pcache2/01629911.jpg'], None, 700, 434, 'Orca spyhopping'], [['https://www.naturepl.com/stock-photo-killer-whales-orcas-orcinus-orca-spyhopping-troms-norway-november-nature-image01629911.html'], 'www.naturepl.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQYPxigtFCbXmkyX5o9az0EF6nYgdTCM0EehMBup_9B19PG88XIPB5gtHrBhMO9ljT81y1IqC_0xOaYx2_yhduPCQbRsX7CUIHfkw'], '[Image of Orca spyhopping]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcRWg59-svCyOoPi-kvWOS2JQ_qdFmYmJ2HlPJyty4wr1J0rWUSeXVPleDRG3wxx'], None, 250, 200], 0, None, None, ['[Image of Orca spyhopping]', 0, 'Orca spyhopping', None, None, 1]]], None, True, None, None, 'en', None, None, [None, [[[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca jumping out of water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca jumping out of water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca jumping out of water]', 0, 'Orca jumping out of water', None, None, 1]], [[['https://capturetheatlas.com/wp-content/uploads/2022/12/Orca-wave-bait-ball.jpg'], None, 900, 600, 'Orca swimming underwater'], [['https://capturetheatlas.com/swimming-with-orcas/'], 'capturetheatlas.com', 5, 'https://encrypted-tbn2.gstatic.com/favicon-tbn?q=tbn:ANd9GcSmt003l7J2VaqnZKIcSAyP5VyMvwr1_YZRA3kyH2Pl8X6uNC6lqvM-CduQSAJmaesYVLFQhX2cX3dnCHQM78ww8EqbNMImpwS-ReAqgw'], '[Image of Orca swimming underwater]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcThgtfk8FZsRWE_FqTBx1n85uzM0PFsALLrev_ypkpz36SApo9PZkl_fhgMWBFH'], None, 250, 200], 0, None, None, ['[Image of Orca swimming underwater]', 0, 'Orca swimming underwater', None, None, 1]], [[['https://www.mensjournal.com/.image/t_share/MTk4Njg1MDA0MDM5OTg4OTk5/killer-whales-monterey-bay.jpg'], None, 1200, 671, 'Orca pod swimming together'], [['https://www.mensjournal.com/news/pod-30-killer-whales-monterey-bay'], 'www.mensjournal.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcSBy9iRr5b6ly40d-TXvQ9mVhN0XAYWUbqXniDVvEGTXMdBUHpsZkD3bNsg9CQYvJzLsZjHYPLn7KuoFQE7C1b46bJ3eKURI4i6McGsIg'], '[Image of Orca pod swimming together]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcT_zbn3aSj3FAvlBoWAq3Fa-S1ZirOYsyG3daLL5b_VhN0_z7-fdLiTW9LwKQ6r'], None, 250, 200], 0, None, None, ['[Image of Orca pod swimming together]', 0, 'Orca pod swimming together', None, None, 1]], [[['https://i.dailymail.co.uk/1s/2021/05/13/12/42941438-0-image-a-13_1620906509879.jpg'], None, 636, 382, 'Orca calf swimming with its mother'], [['https://www.dailymail.co.uk/sciencetech/article-9574917/Orca-enthusiast-captures-incredible-photo-rare-newborn-calf-coast-Scotland.html'], 'www.dailymail.co.uk', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcTxhjGkCz1VjtErZLwi2Qoupz4JRVjuXvn5-XVV95YtBS9w8GoFJ36mkZqAG33esE0y2aqDVRmnxtwu7CbUzJFR5TGZCQDTQq-fCGJ-kA'], '[Image of Orca calf swimming with its mother]', [['https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSsAMSHd01qPovodn7BzzGKE_u1SpRwMLVrsWo_VexZRGf3PMipYd9gku8uhr78'], None, 250, 200], 0, None, None, ['[Image of Orca calf swimming with its mother]', 0, 'Orca calf swimming with its mother', None, None, 1]], [[['https://nplimages.infradoxxs.com/cache/pcache2/01629911.jpg'], None, 700, 434, 'Orca spyhopping'], [['https://www.naturepl.com/stock-photo-killer-whales-orcas-orcinus-orca-spyhopping-troms-norway-november-nature-image01629911.html'], 'www.naturepl.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQYPxigtFCbXmkyX5o9az0EF6nYgdTCM0EehMBup_9B19PG88XIPB5gtHrBhMO9ljT81y1IqC_0xOaYx2_yhduPCQbRsX7CUIHfkw'], '[Image of Orca spyhopping]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcRWg59-svCyOoPi-kvWOS2JQ_qdFmYmJ2HlPJyty4wr1J0rWUSeXVPleDRG3wxx'], None, 250, 200], 0, None, None, ['[Image of Orca spyhopping]', 0, 'Orca spyhopping', None, None, 1]]], None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], [1], None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False], ['rc_85b578db36624048', ["Here are 5 pictures of orcas:\n\n**[1] A family of orcas swimming in the wild.**\n[Image of Orca whale family swimming]\n\n**[2] An orca jumping out of the water.**\n[Image of Orca whale jumping out of the water]\n\n**[3] An orca swimming underwater.**\n[Image of Orca whale swimming underwater]\n\n**[4] An orca with its calf.**\n[Image of Orca whale with its calf]\n\n**[5] A close-up of an orca's face.**\n[Image of Orca whale close up]\n\nOrcas are apex predators and have a diverse diet. Individual populations often specialize in particular types of prey. This includes a variety of fish, sharks (including the great white shark), rays, and marine mammals such as seals, other species of dolphin, and whales.\n\nOrcas are highly intelligent social animals and live in pods of up to 50 individuals. Pods are typically led by a dominant female and consist of related females and their offspring. Males leave their natal pod to join other pods, and they may travel long distances in search of mates.\n\nOrcas are an important part of the marine ecosystem and play a vital role in keeping populations of other marine animals in check. They are also a popular tourist attraction, and many people enjoy watching them in their natural habitat."], [[[426, 696, ['https://en.wikipedia.org/wiki/Orca', '', '', '', '', None, None, '', False, '', '', '', '', '', '', ''], 2, 2, None, [1, 'normal_citation_datasets']]]], None, [[[['https://ih1.redbubble.net/image.1584986020.2098/flat,750x,075,f-pad,750x1000,f8f8f8.jpg'], None, 750, 1000, 'Orca whale family swimming'], [['https://www.redbubble.com/i/art-board-print/Orca-family-swimming-underwater-by-Eleanarte/56002098.8DB3C'], 'www.redbubble.com', 5, 'https://encrypted-tbn1.gstatic.com/favicon-tbn?q=tbn:ANd9GcSt7MDDralvq7Cus2iaFC8JpfABRui6gXpNFnr3cjjqjJjlEWS6fvwnRp-_wvje1k6NNtyRVrQrj8H3390CP18PqM9oah8QkIZvPtw'], '[Image of Orca whale family swimming]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSn8SUctkmWV8_dTEZT13lrz4kdX7dGsinI822B8-gbw-p4tlFg92ggGRgUKgz0'], None, 250, 200], 0, None, None, ['[Image of Orca whale family swimming]', 0, 'Orca whale family swimming', None, None, 1]], [[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca whale jumping out of the water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca whale jumping out of the water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca whale jumping out of the water]', 0, 'Orca whale jumping out of the water', None, None, 1]], [[['https://capturetheatlas.com/wp-content/uploads/2022/12/Orca-wave-bait-ball.jpg'], None, 900, 600, 'Orca whale swimming underwater'], [['https://capturetheatlas.com/swimming-with-orcas/'], 'capturetheatlas.com', 5, 'https://encrypted-tbn2.gstatic.com/favicon-tbn?q=tbn:ANd9GcSmt003l7J2VaqnZKIcSAyP5VyMvwr1_YZRA3kyH2Pl8X6uNC6lqvM-CduQSAJmaesYVLFQhX2cX3dnCHQM78ww8EqbNMImpwS-ReAqgw'], '[Image of Orca whale swimming underwater]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcThgtfk8FZsRWE_FqTBx1n85uzM0PFsALLrev_ypkpz36SApo9PZkl_fhgMWBFH'], None, 250, 200], 0, None, None, ['[Image of Orca whale swimming underwater]', 0, 'Orca whale swimming underwater', None, None, 1]], [[['https://media.cnn.com/api/v1/images/stellar/prod/200907041152-killer-whale-new-baby.jpg?q=w_1600,h_1016,x_0,y_0,c_fill'], None, 1600, 1016, 'Orca whale with its calf'], [['https://www.cnn.com/2020/09/07/us/killer-whale-new-calf-scn-trnd/index.html'], 'www.cnn.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcR7W_OzfcVuPH1nzycyRRtqBT-dc7A2PWZ7o1msLOBuX45_l8a-_mUJdSzlIyrBZe-IvXw4HzwZhospc5scaQxtbGSDgwk'], '[Image of Orca whale with its calf]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcQV1jS9B72wq8BLSm6_3HBseZGqDCeMbmXxNXYGkSEC_Bai6G5xuIn9V8r9O7IB'], None, 250, 200], 0, None, None, ['[Image of Orca whale with its calf]', 0, 'Orca whale with its calf', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/tours/killer-whale-up-close/2-column/killer-whale-up-close-tour_500x375.ashx?version=1_202007284351&mw=400&mh=400&as=0&hash=1280E82513A9D1792DE55BBC282D94C43A2B3782'], None, 400, 300, 'Orca whale close up'], [['https://seaworld.com/orlando/tours/killer-whale-up-close-tour/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca whale close up]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSoXxivDwDkgPLAKZKRNK4GvTsYo_zcw5qGDtgsdSnMOl_vMdaN6_-6LokjZGjc'], None, 250, 200], 0, None, None, ['[Image of Orca whale close up]', 0, 'Orca whale close up', None, None, 1]]], None, True, None, None, 'en', None, None, [None, [[[['https://ih1.redbubble.net/image.1584986020.2098/flat,750x,075,f-pad,750x1000,f8f8f8.jpg'], None, 750, 1000, 'Orca whale family swimming'], [['https://www.redbubble.com/i/art-board-print/Orca-family-swimming-underwater-by-Eleanarte/56002098.8DB3C'], 'www.redbubble.com', 5, 'https://encrypted-tbn1.gstatic.com/favicon-tbn?q=tbn:ANd9GcSt7MDDralvq7Cus2iaFC8JpfABRui6gXpNFnr3cjjqjJjlEWS6fvwnRp-_wvje1k6NNtyRVrQrj8H3390CP18PqM9oah8QkIZvPtw'], '[Image of Orca whale family swimming]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSn8SUctkmWV8_dTEZT13lrz4kdX7dGsinI822B8-gbw-p4tlFg92ggGRgUKgz0'], None, 250, 200], 0, None, None, ['[Image of Orca whale family swimming]', 0, 'Orca whale family swimming', None, None, 1]], [[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca whale jumping out of the water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca whale jumping out of the water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca whale jumping out of the water]', 0, 'Orca whale jumping out of the water', None, None, 1]], [[['https://capturetheatlas.com/wp-content/uploads/2022/12/Orca-wave-bait-ball.jpg'], None, 900, 600, 'Orca whale swimming underwater'], [['https://capturetheatlas.com/swimming-with-orcas/'], 'capturetheatlas.com', 5, 'https://encrypted-tbn2.gstatic.com/favicon-tbn?q=tbn:ANd9GcSmt003l7J2VaqnZKIcSAyP5VyMvwr1_YZRA3kyH2Pl8X6uNC6lqvM-CduQSAJmaesYVLFQhX2cX3dnCHQM78ww8EqbNMImpwS-ReAqgw'], '[Image of Orca whale swimming underwater]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcThgtfk8FZsRWE_FqTBx1n85uzM0PFsALLrev_ypkpz36SApo9PZkl_fhgMWBFH'], None, 250, 200], 0, None, None, ['[Image of Orca whale swimming underwater]', 0, 'Orca whale swimming underwater', None, None, 1]], [[['https://media.cnn.com/api/v1/images/stellar/prod/200907041152-killer-whale-new-baby.jpg?q=w_1600,h_1016,x_0,y_0,c_fill'], None, 1600, 1016, 'Orca whale with its calf'], [['https://www.cnn.com/2020/09/07/us/killer-whale-new-calf-scn-trnd/index.html'], 'www.cnn.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcR7W_OzfcVuPH1nzycyRRtqBT-dc7A2PWZ7o1msLOBuX45_l8a-_mUJdSzlIyrBZe-IvXw4HzwZhospc5scaQxtbGSDgwk'], '[Image of Orca whale with its calf]', [['https://encrypted-tbn2.gstatic.com/images?q=tbn:ANd9GcQV1jS9B72wq8BLSm6_3HBseZGqDCeMbmXxNXYGkSEC_Bai6G5xuIn9V8r9O7IB'], None, 250, 200], 0, None, None, ['[Image of Orca whale with its calf]', 0, 'Orca whale with its calf', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/tours/killer-whale-up-close/2-column/killer-whale-up-close-tour_500x375.ashx?version=1_202007284351&mw=400&mh=400&as=0&hash=1280E82513A9D1792DE55BBC282D94C43A2B3782'], None, 400, 300, 'Orca whale close up'], [['https://seaworld.com/orlando/tours/killer-whale-up-close-tour/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca whale close up]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSoXxivDwDkgPLAKZKRNK4GvTsYo_zcw5qGDtgsdSnMOl_vMdaN6_-6LokjZGjc'], None, 250, 200], 0, None, None, ['[Image of Orca whale close up]', 0, 'Orca whale close up', None, None, 1]]], None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], [1], None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False], ['rc_85b578db366245c3', ["Here are 5 pictures of orcas:\n\n1. **Orca jumping out of the water:**\n[Image of Orca jumping out of the water]\n2. **Orca family swimming together:**\n[Image of Orca family swimming together]\n3. **Orca breaching the surface:**\n[Image of Orca breaching the surface]\n4. **Orca underwater:**\n[Image of Orca underwater]\n5. **Orca close-up:**\n[Image of Orca close-up]\n\nOrcas are the largest members of the dolphin family and are found in all of the world's oceans. They are intelligent and social animals, and live in groups called pods. Orcas are apex predators, meaning they are at the top of the food chain. They have a diverse diet, including fish, sharks, seals, and other whales.\n\nOrcas are fascinating creatures, and these pictures capture some of their beauty and power."], [], None, [[[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca jumping out of the water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca jumping out of the water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca jumping out of the water]', 0, 'Orca jumping out of the water', None, None, 1]], [[['https://www.lowerwoodlandstudio.com/cdn/shop/products/Orca-Family_1000x.jpg?v=1581541431'], None, 1000, 1000, 'Orca family swimming together'], [['https://www.lowerwoodlandstudio.com/products/orca-family-art-print'], 'www.lowerwoodlandstudio.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcTrq_UvMkvGvKQx7oBRE24aUpbk9PXzwuQby67S_xzkofp1rNAptrzfWoRlhy8GKEU7-vElIWznGHLAVEU1r2bXiLKaxDLUSEVb9vDm1ZLwQrDe6mS9'], '[Image of Orca family swimming together]', [['https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS5sHfBv2xqxNgR5mZRtSmanu691M0FuTcfnygqxHjhkNLvP7IRgHg47bC-tWOc'], None, 250, 200], 0, None, None, ['[Image of Orca family swimming together]', 0, 'Orca family swimming together', None, None, 1]], [[['https://d3ftabzjnxfdg6.cloudfront.net/app/uploads/2018/11/slide2.jpg'], None, 1023, 630, 'Orca breaching the surface'], [['https://www.eaglewingtours.com/'], 'www.eaglewingtours.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcR6OdEyW_bcmB4JAPfwSkcBrXK7fwOAOFGI2cak0Pmc7Ouphw2qoGeJWL7x8mK7I74C4LqlMO60uvyr_00a9hRasjaXkJAujcMU_96uua65Hw'], '[Image of Orca breaching the surface]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSjxtPQLbcLmK2sou4gFq0Da8WVqQMJ74xqNQ50HWk6kCnQ-44HRG-LMNfvBxiQ'], None, 250, 200], 0, None, None, ['[Image of Orca breaching the surface]', 0, 'Orca breaching the surface', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/featured-story/get-closer-to-killer-whales-at-seaworld.ashx'], None, 660, 410, 'Orca underwater'], [['https://seaworld.com/orlando/animal-experiences/orca-underwater-viewing/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca underwater]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSU0YIW5dRVc_n6xIV2SuIIgfQDmSo9TmhRnmB-7Mo33ZoaiHpsVtfAIZp68ztX'], None, 250, 200], 0, None, None, ['[Image of Orca underwater]', 0, 'Orca underwater', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/tours/killer-whale-up-close/2-column/killer-whale-up-close-tour_500x375.ashx?version=1_202007284351&mw=400&mh=400&as=0&hash=1280E82513A9D1792DE55BBC282D94C43A2B3782'], None, 400, 300, 'Orca close-up'], [['https://seaworld.com/orlando/tours/killer-whale-up-close-tour/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca close-up]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSoXxivDwDkgPLAKZKRNK4GvTsYo_zcw5qGDtgsdSnMOl_vMdaN6_-6LokjZGjc'], None, 250, 200], 0, None, None, ['[Image of Orca close-up]', 0, 'Orca close-up', None, None, 1]]], None, True, None, None, 'en', None, None, [None, [[[['https://cdn.outsideonline.com/wp-content/uploads/2016/03/17/seaworld-captivity-breeding_s.jpg'], None, 1500, 1500, 'Orca jumping out of the water'], [['https://www.outsideonline.com/outdoor-adventure/water-activities/seaworld-end-killer-whale-shows/'], 'www.outsideonline.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcRnI9jjsNbOPGuQTOXyAZMa9-MW4A8Y4qCOUwP-HqyzOFONSabVNw78H3-FStoW-gEL42WkDW9q0eNtBq11M6F6gRmYaareDSSQV3m6bSNY'], '[Image of Orca jumping out of the water]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSasL8NXAbsNgAcc_vKCJPLW2GyZk8d6fWmQ3inAIt8Nao4-i8tnm_wyTsTXCNW'], None, 250, 200], 0, None, None, ['[Image of Orca jumping out of the water]', 0, 'Orca jumping out of the water', None, None, 1]], [[['https://www.lowerwoodlandstudio.com/cdn/shop/products/Orca-Family_1000x.jpg?v=1581541431'], None, 1000, 1000, 'Orca family swimming together'], [['https://www.lowerwoodlandstudio.com/products/orca-family-art-print'], 'www.lowerwoodlandstudio.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcTrq_UvMkvGvKQx7oBRE24aUpbk9PXzwuQby67S_xzkofp1rNAptrzfWoRlhy8GKEU7-vElIWznGHLAVEU1r2bXiLKaxDLUSEVb9vDm1ZLwQrDe6mS9'], '[Image of Orca family swimming together]', [['https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS5sHfBv2xqxNgR5mZRtSmanu691M0FuTcfnygqxHjhkNLvP7IRgHg47bC-tWOc'], None, 250, 200], 0, None, None, ['[Image of Orca family swimming together]', 0, 'Orca family swimming together', None, None, 1]], [[['https://d3ftabzjnxfdg6.cloudfront.net/app/uploads/2018/11/slide2.jpg'], None, 1023, 630, 'Orca breaching the surface'], [['https://www.eaglewingtours.com/'], 'www.eaglewingtours.com', 5, 'https://encrypted-tbn3.gstatic.com/favicon-tbn?q=tbn:ANd9GcR6OdEyW_bcmB4JAPfwSkcBrXK7fwOAOFGI2cak0Pmc7Ouphw2qoGeJWL7x8mK7I74C4LqlMO60uvyr_00a9hRasjaXkJAujcMU_96uua65Hw'], '[Image of Orca breaching the surface]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSjxtPQLbcLmK2sou4gFq0Da8WVqQMJ74xqNQ50HWk6kCnQ-44HRG-LMNfvBxiQ'], None, 250, 200], 0, None, None, ['[Image of Orca breaching the surface]', 0, 'Orca breaching the surface', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/featured-story/get-closer-to-killer-whales-at-seaworld.ashx'], None, 660, 410, 'Orca underwater'], [['https://seaworld.com/orlando/animal-experiences/orca-underwater-viewing/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca underwater]', [['https://encrypted-tbn1.gstatic.com/images?q=tbn:ANd9GcSU0YIW5dRVc_n6xIV2SuIIgfQDmSo9TmhRnmB-7Mo33ZoaiHpsVtfAIZp68ztX'], None, 250, 200], 0, None, None, ['[Image of Orca underwater]', 0, 'Orca underwater', None, None, 1]], [[['https://seaworld.com/orlando/-/media/seaworld-orlando/images/tours/killer-whale-up-close/2-column/killer-whale-up-close-tour_500x375.ashx?version=1_202007284351&mw=400&mh=400&as=0&hash=1280E82513A9D1792DE55BBC282D94C43A2B3782'], None, 400, 300, 'Orca close-up'], [['https://seaworld.com/orlando/tours/killer-whale-up-close-tour/'], 'seaworld.com', 5, 'https://encrypted-tbn0.gstatic.com/favicon-tbn?q=tbn:ANd9GcQgtzOgKyfUom1chEpEkvb3qFmEKVC3MBKNy6zJ2UwQ-O0dgLlq_9P3rgzRs0b8LrC-CKmosXeySAP8885RzZRZubauJfz8'], '[Image of Orca close-up]', [['https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSoXxivDwDkgPLAKZKRNK4GvTsYo_zcw5qGDtgsdSnMOl_vMdaN6_-6LokjZGjc'], None, 250, 200], 0, None, None, ['[Image of Orca close-up]', 0, 'Orca close-up', None, None, 1]]], None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], [1], None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False]], ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599104211, 107296609], [599287085, 107661471]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1147.2634, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=YYOhkA_Esek-Ta587n_qx_oy03StDDDWOMCuB-AWD8QAVPYaJKCVuADaeBSzQDNgvMTmKEbx6pwVt1oMIL2C8tU0Avp0wGh6xbqIoqtO1ON2wHjwPneZhJusVyEjOa_1JFJOyvDMLn1IjFC3QO7XRhKle1BV4wodoVUAc9asy-oro6ybx8IU3kOjxmONIe8M'], None, None, 'NO', []], [None, ['c_85b578db36624ef5', 'r_85b578db36624470'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599104211, 107296609], [599287085, 107661471]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1147.2634, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=YYOhkA_Esek-Ta587n_qx_oy03StDDDWOMCuB-AWD8QAVPYaJKCVuADaeBSzQDNgvMTmKEbx6pwVt1oMIL2C8tU0Avp0wGh6xbqIoqtO1ON2wHjwPneZhJusVyEjOa_1JFJOyvDMLn1IjFC3QO7XRhKle1BV4wodoVUAc9asy-oro6ybx8IU3kOjxmONIe8M'], None, None, 'NO', None, ['Orcas in the wild']]]
 
 # Show 2 pictures of orcas
@@ -15,3 +17,9 @@ t5 = [[None, ['c_298eb3e0345e6509', 'r_298eb3e0345e6478'], None, None, None, ['O
 
 # Show me flights to visit my aunt in Madrid the week of January 22. Give me gift ideas for what to bring her, she loves dolphins and playing cards
 t6 = [[None, ['c_61a4e9e841399fcf', 'r_61a4e9e84139919a'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599108562, 107305290], [599282734, 107652790]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1092.671, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=o4Pr0NR4wYKvuqtv6a-Cb9uEMiPTSPlC-LUxUXhnd6e6v4Mt-NKmsZKk-TuOwlEprIWShY6yiVxUSVUrqNZF0OipBKOHtYlX0nNHcmv8FM-HxZciNgFcUaRv7nQvk6rsKC0St4qTnEBbdNy2Qa3A28BkYRtChWq2BVpBHNCnbOnDFg7ojkWjwA0iHqukBVTM'], [None, ['google_flights_tool', [None, None, 'Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg'], 1]], None, 'NO', []], [None, ['c_61a4e9e841399fcf', 'r_61a4e9e84139919a'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599108562, 107305290], [599282734, 107652790]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1092.671, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=o4Pr0NR4wYKvuqtv6a-Cb9uEMiPTSPlC-LUxUXhnd6e6v4Mt-NKmsZKk-TuOwlEprIWShY6yiVxUSVUrqNZF0OipBKOHtYlX0nNHcmv8FM-HxZciNgFcUaRv7nQvk6rsKC0St4qTnEBbdNy2Qa3A28BkYRtChWq2BVpBHNCnbOnDFg7ojkWjwA0iHqukBVTM'], [None, ['google_flights_tool', ['', None, 'Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg'], 2]], None, 'NO', []], [None, ['c_61a4e9e841399fcf', 'r_61a4e9e84139919a'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599108562, 107305290], [599282734, 107652790]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1092.671, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=o4Pr0NR4wYKvuqtv6a-Cb9uEMiPTSPlC-LUxUXhnd6e6v4Mt-NKmsZKk-TuOwlEprIWShY6yiVxUSVUrqNZF0OipBKOHtYlX0nNHcmv8FM-HxZciNgFcUaRv7nQvk6rsKC0St4qTnEBbdNy2Qa3A28BkYRtChWq2BVpBHNCnbOnDFg7ojkWjwA0iHqukBVTM'], [None, ['google_flights_tool', [None, '', 'Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg'], 3]], None, 'NO', []], [None, ['c_61a4e9e841399fcf', 'r_61a4e9e84139919a'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599108562, 107305290], [599282734, 107652790]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1092.671, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=o4Pr0NR4wYKvuqtv6a-Cb9uEMiPTSPlC-LUxUXhnd6e6v4Mt-NKmsZKk-TuOwlEprIWShY6yiVxUSVUrqNZF0OipBKOHtYlX0nNHcmv8FM-HxZciNgFcUaRv7nQvk6rsKC0St4qTnEBbdNy2Qa3A28BkYRtChWq2BVpBHNCnbOnDFg7ojkWjwA0iHqukBVTM'], [None, ['google_flights_tool', [None, None, 'Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg'], 4]], None, 'NO', []], [None, ['c_61a4e9e841399fcf', 'r_61a4e9e84139919a'], [['gift ideas for dolphin lover who plays cards', 1], ['What to get someone who likes to make cards?', 4], ['What are some priceless gifts?', 4], ['What do dolphins like to play with?', 4], ['How do dolphins express love?', 4]], None, [['rc_61a4e9e841399e27', ['http://googleusercontent.com/tool_disclaimer_content/1 I found a few round trip flights from Oslo (OSL, TRF) to Madrid (MAD, XOC, XTI) departing on January 22 and returning on January 28.\n\nHere are some of the departing flights from Oslo:\n  * Iberia IB3497 departing at 15:00, arriving at 18:50, priced from NOK 1,595 (round trip) will take 3 hours and 50 minutes.\n  * Air France AF1775 AF1400 departing at 17:20, arriving at 23:05, priced from NOK 1,766 (round trip) will take 5 hours and 45 minutes.\n  * KLM KL1142 KL1703 departing at 09:20, arriving at 16:05, priced from NOK 1,885 (round trip) will take 6 hours and 45 minutes.\n\nYou can find more details on [Google Flights](https://www.google.com/travel/flights?tfs=CCcQAhooEgoyMDI0LTAxLTIyagwIAxIIL20vMDVsNjRyDAgDEggvbS8wNTZfeRooEgoyMDI0LTAxLTI4agwIAxIIL20vMDU2X3lyDAgDEggvbS8wNWw2NFIDTk9LemxDalJJTkdsaVZsUmxOVk5tTVRoQlJHZHBOa0ZDUnkwdExTMHRMUzB0YkdaaVkyWXpOa0ZCUVVGQlIxVlJjakpyUVdOdFVYVkJFZ1pKUWpNME9UY2FDZ2k3REJBQUdnTk9UMHM0TTNEbGN3PT0)\n\nHere are some gift ideas for your aunt who loves dolphins and playing cards:\n\n* A dolphin-themed gift basket filled with her favorite snacks and treats.\n* A personalized dolphin mug or water bottle.\n* A dolphin-themed piece of jewelry, such as a necklace or bracelet.\n* A dolphin-themed painting or print for her home.\n* A deck of dolphin-themed playing cards.\n* A gift certificate to a local aquarium or marine park.\n* A donation to her favorite dolphin conservation charity in her name.\n\nI hope these ideas help you find the perfect gift for your aunt!\nhttp://googleusercontent.com/flight_content/0'], [], None, None, None, False, None, None, 'en', None, None, [None, None, None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 4], None, None, None, None, None, None, None, None, False]], None, None, None, None, None, None, None, None, None, None, [[2, [[[['Iberia'], 'https://www.gstatic.com/flights/airline_logos/70px/IB.png', 'OSL', 'MAD', None, None, 0, '15:00', '18:50', '3 hr 50 min'], None, 'https://www.google.com/travel/flights?tfs=CCcQAhpKEgpKYW51YXJ5IDIyIiAKA09TTBIKMjAyNC0wMS0yMhoDTUFEKgJJQjIEMzQ5N2oMCAMSCC9tLzA1bDY0cgwIAxIIL20vMDU2X3kaShIKSmFudWFyeSAyOCIgCgNPU0wSCjIwMjQtMDEtMjIaA01BRCoCSUIyBDM0OTdqDAgDEggvbS8wNTZfeXIMCAMSCC9tLzA1bDY0UgNOT0t6bENqUklOR2xpVmxSbE5WTm1NVGhCUkdkcE5rRkNSeTB0TFMwdExTMHRiR1ppWTJZek5rRkJRVUZCUjFWUmNqSnJRV050VVhWQkVnWkpRak0wT1RjYUNnaTdEQkFBR2dOT1QwczRNM0RsY3c9PQ', 'NOK\xa01,595'], [[['Air France'], 'https://www.gstatic.com/flights/airline_logos/70px/AF.png', 'OSL', 'MAD', None, None, 1, '17:20', '23:05', '5 hr 45 min'], None, 'https://www.google.com/travel/flights?tfs=CCcQAhpsEgpKYW51YXJ5IDIyIiAKA09TTBIKMjAyNC0wMS0yMhoDQ0RHKgJBRjIEMTc3NSIgCgNDREcSCjIwMjQtMDEtMjIaA01BRCoCQUYyBDE0MDBqDAgDEggvbS8wNWw2NHIMCAMSCC9tLzA1Nl95GmwSCkphbnVhcnkgMjgiIAoDT1NMEgoyMDI0LTAxLTIyGgNDREcqAkFGMgQxNzc1IiAKA0NERxIKMjAyNC0wMS0yMhoDTUFEKgJBRjIEMTQwMGoMCAMSCC9tLzA1Nl95cgwIAxIIL20vMDVsNjRSA05PS3p0Q2pSSU5HbGlWbFJsTlZObU1UaEJSR2RwTmtGQ1J5MHRMUzB0TFMwdGJHWmlZMll6TmtGQlFVRkJSMVZSY2pKclFXTnRVWFZCRWcxQlJqRTNOelY4UVVZeE5EQXdHZ29JNWcwUUFCb0RUazlMT0ROd21vQUI', 'NOK\xa01,766'], [[['KLM'], 'https://www.gstatic.com/flights/airline_logos/70px/KL.png', 'OSL', 'MAD', None, None, 1, '09:20', '16:05', '6 hr 45 min'], None, 'https://www.google.com/travel/flights?tfs=CCcQAhpsEgpKYW51YXJ5IDIyIiAKA09TTBIKMjAyNC0wMS0yMhoDQU1TKgJLTDIEMTE0MiIgCgNBTVMSCjIwMjQtMDEtMjIaA01BRCoCS0wyBDE3MDNqDAgDEggvbS8wNWw2NHIMCAMSCC9tLzA1Nl95GmwSCkphbnVhcnkgMjgiIAoDT1NMEgoyMDI0LTAxLTIyGgNBTVMqAktMMgQxMTQyIiAKA0FNUxIKMjAyNC0wMS0yMhoDTUFEKgJLTDIEMTcwM2oMCAMSCC9tLzA1Nl95cgwIAxIIL20vMDVsNjRSA05PS3p0Q2pSSU5HbGlWbFJsTlZObU1UaEJSR2RwTmtGQ1J5MHRMUzB0TFMwdGJHWmlZMll6TmtGQlFVRkJSMVZSY2pKclFXTnRVWFZCRWcxTFRERXhOREo4UzB3eE56QXpHZ29JM1E0UUFCb0RUazlMT0ROdzY0Z0I', 'NOK\xa01,885'], [[['KLM'], 'https://www.gstatic.com/flights/airline_logos/70px/KL.png', 'TRF', 'MAD', None, None, 1, '18:10', '23:25', '5 hr 15 min'], None, 'https://www.google.com/travel/flights?tfs=CCcQAhpsEgpKYW51YXJ5IDIyIiAKA1RSRhIKMjAyNC0wMS0yMhoDQU1TKgJLTDIEMTIxOCIgCgNBTVMSCjIwMjQtMDEtMjIaA01BRCoCS0wyBDE3MDdqDAgDEggvbS8wNWw2NHIMCAMSCC9tLzA1Nl95GmwSCkphbnVhcnkgMjgiIAoDVFJGEgoyMDI0LTAxLTIyGgNBTVMqAktMMgQxMjE4IiAKA0FNUxIKMjAyNC0wMS0yMhoDTUFEKgJLTDIEMTcwN2oMCAMSCC9tLzA1Nl95cgwIAxIIL20vMDVsNjRSA05PS3p0Q2pSSU5HbGlWbFJsTlZObU1UaEJSR2RwTmtGQ1J5MHRMUzB0TFMwdGJHWmlZMll6TmtGQlFVRkJSMVZSY2pKclFXTnRVWFZCRWcxTFRERXlNVGg4UzB3eE56QTNHZ29JL2d3UUFCb0RUazlMT0ROdzAzZz0', 'NOK\xa01,662'], [[['SWISS'], 'https://www.gstatic.com/flights/airline_logos/70px/LX.png', 'OSL', 'MAD', None, None, 1, '08:40', '19:00', '10 hr 20 min'], None, 'https://www.google.com/travel/flights?tfs=CCcQAhpsEgpKYW51YXJ5IDIyIiAKA09TTBIKMjAyNC0wMS0yMhoDWlJIKgJMWDIEMTIxOSIgCgNaUkgSCjIwMjQtMDEtMjIaA01BRCoCTFgyBDIwMzJqDAgDEggvbS8wNWw2NHIMCAMSCC9tLzA1Nl95GmwSCkphbnVhcnkgMjgiIAoDT1NMEgoyMDI0LTAxLTIyGgNaUkgqAkxYMgQxMjE5IiAKA1pSSBIKMjAyNC0wMS0yMhoDTUFEKgJMWDIEMjAzMmoMCAMSCC9tLzA1Nl95cgwIAxIIL20vMDVsNjRSA05PS3p0Q2pSSU5HbGlWbFJsTlZObU1UaEJSR2RwTmtGQ1J5MHRMUzB0TFMwdGJHWmlZMll6TmtGQlFVRkJSMVZSY2pKclFXTnRVWFZCRWcxTVdERXlNVGw4VEZneU1ETXlHZ29JaXcwUUFCb0RUazlMT0ROd3pIaz0', 'NOK\xa01,675']], 'https://www.google.com/travel/flights?tfs=CCcQAhooEgoyMDI0LTAxLTIyagwIAxIIL20vMDVsNjRyDAgDEggvbS8wNTZfeRooEgoyMDI0LTAxLTI4agwIAxIIL20vMDU2X3lyDAgDEggvbS8wNWw2NFIDTk9LemxDalJJTkdsaVZsUmxOVk5tTVRoQlJHZHBOa0ZDUnkwdExTMHRMUzB0YkdaaVkyWXpOa0ZCUVVGQlIxVlJjakpyUVdOdFVYVkJFZ1pKUWpNME9UY2FDZ2k3REJBQUdnTk9UMHM0TTNEbGN3PT0', ['http://googleusercontent.com/flight_content/0', None, 'Round trip · Oslo (OSL, TRF) - Madrid (MAD, XOC + 1 more) · Jan 22 - Jan 28 · 1 adult', None, None, None, ['Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg']], 'Oslo (OSL, TRF)', 'Madrid (MAD, XOC + 1 more)', 'Jan 22', 'Jan 28', '1 adult']], None, None, [[None, None, None, None, None, None, None, False]], None, None, [[['http://googleusercontent.com/tool_disclaimer_content/1'], "Bard doesn't always get it right. Be sure to check the following flight details:\n\n*Round trip* · *Oslo (OSL, TRF)* - Madrid (MAD, XOC + 1 more) · Jan 22 - *Jan 28* · *1 adult*"]]], None, False, False, False, None, None, False]], ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599108562, 107305290], [599282734, 107652790]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1092.671, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=o4Pr0NR4wYKvuqtv6a-Cb9uEMiPTSPlC-LUxUXhnd6e6v4Mt-NKmsZKk-TuOwlEprIWShY6yiVxUSVUrqNZF0OipBKOHtYlX0nNHcmv8FM-HxZciNgFcUaRv7nQvk6rsKC0St4qTnEBbdNy2Qa3A28BkYRtChWq2BVpBHNCnbOnDFg7ojkWjwA0iHqukBVTM'], None, None, 'NO', [], None, [[None, ['google_flights_tool', [None, None, 'Google Flights', 'https://www.gstatic.com/images/branding/productlogos/travel_flights/v7/192px.svg'], 4]]]]]
+
+# What is the difference between __repr__ and __str__ in python?
+t7 = [[None, ['c_903e0fc7657e7d7a', 'r_903e0fc7657e7dd7'], [['difference between __repr__ and __str__ in python', 1], ['What is the use of __str __ in Python?', 4], ['What is the difference between __ new __ and __ init __?', 4]], None, [['rc_903e0fc7657e7062', ['The two special methods `__repr__()` and `__str__()` are both used to return a string representation of an object in Python. However, they have different purposes.\n\n**`__repr__()`** is used to return the **official** string representation of an object. This representation should be unambiguous and unambiguous, and should be able to be used to recreate the object. It is intended for developers, and is often used for debugging and logging purposes.\n\n**`__str__()`** is used to return a more **informal** string representation of an object. It is intended for users, and should be easy to read and understand.\n\nIf you do not define a `__str__()` method for your class, the built-in object implementation will call the `__repr__()` method instead.\n\nHere is an example:\n\n```python\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n```\n\nOutput:\n\n```\nA\nA str\n```\n\nIn this example, the `__repr__()` method simply returns the class name, while the `__str__()` method returns a more descriptive string.\n\nWhen to use `__repr__()` and `__str__()`\n\n* Use `__repr__()` when you need to get a unambiguous and unambiguous string representation of an object for debugging or logging purposes.\n* Use `__str__()` when you need to get a human-readable string representation of an object for displaying to the user.\n\nHere are some additional tips for using `__repr__()` and `__str__()`:\n\n* The `__repr__()` method should always return a string that is valid Python code. This means that you should be able to evaluate the string in Python and get the original object back.\n* The `__str__()` method should not return a string that is valid Python code. This is because `__str__()` is intended for users, and users should not have to worry about the internal details of your class.\n* If you are defining a `__repr__()` method for a class that contains other objects, you should use the `__repr__()` method of those objects to generate the string representation. This will ensure that the string representation is complete and accurate.\n\nI hope this helps!'], [], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, 'A\nA str\n', '', 'what is the difference between __repr__ and __str__ in python?', None, 0, [], 1, '\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', None, None, None, None, None, None, None, 3, None, None, None, None, None, [None, None, True, None, None, None, None, None, None, None, None, None, 1, False, False], None, None, None, None, None, None, None, None, True]], None, None, None, None, None, None, None, None, None, None, None, None, None, [['\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', 0, 3, None, None, None, None, True]]], None, False, True, True, None, None, False], ['rc_903e0fc7657e70bf', ['The `__repr__()` and `__str__()` methods are both special methods in Python that are used to return a string representation of an object. However, there are some key differences between the two:\n\n* **__repr__()** is intended to provide a unique and unambiguous representation of the object, that can be used to recreate the object later. It is often used for debugging and logging purposes.\n* **__str__()** is intended to provide a human-readable representation of the object. It is often used for displaying the object to the user, or for saving the object to a file.\n\nIn general, it is a good practice to define both `__repr__()` and `__str__()` for your classes. If you only define one, then it is generally recommended to define `__repr__()`, as this can be used to generate a `__str__()` implementation if needed.\n\nHere is an example of a simple class that defines both `__repr__()` and `__str__()`:\n\n```python\nclass Person:\n    def __init__(self, name, age):\n        self.name = name\n        self.age = age\n\n    def __repr__(self):\n        return f"Person(\'{self.name}\', {self.age})"\n\n    def __str__(self):\n        return f"{self.name} is {self.age} years old."\n\n```\n\nIf we create an instance of this class and print it, we get the following output:\n\n```python\n>>> person = Person(\'Alice\', 25)\n>>> print(person)\nAlice is 25 years old.\n```\n\nHowever, if we use the `repr()` function to print the object, we get a different output:\n\n```python\n>>> repr(person)\n"Person(\'Alice\', 25)"\n```\n\nThis output is more verbose, but it also contains all of the information that is needed to recreate the object.\n\nIn general, it is best to use the `str()` function to print objects to the user, and the `repr()` function to print objects for debugging or logging purposes.'], [[[909, 1095, ['', 'edelahoz/Introduction-to-Python', '', '', '', None, None, '', False, '', '', '', '', '', '', ''], 1, 100, None, [1]], [940, 1139, ['', 'Engr-Asad-Hussain/oop', '', '', '', None, None, '', False, '', '', '', '', '', '', ''], 1, 100, None, [1]], [953, 1166, ['', 'TeknikhogskolanGothenburg/PGBPYH21_Programmering', '', '', '', None, None, '', False, '', '', '', '', '', '', ''], 1, 100, None, [1]]]], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, 'A\nA str\n', '', 'what is the difference between __repr__ and __str__ in python?', None, 0, [], 1, '\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', None, None, None, None, None, None, None, 3, None, None, None, None, None, [None, None, True, None, None, None, None, None, None, None, None, None, 1, False, False], None, None, None, None, None, None, None, None, True]], None, None, None, None, None, None, None, None, None, None, None, None, None, [['\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', 0, 3, None, None, None, None, True]]], None, False, True, True, None, None, False], ['rc_903e0fc7657e711c', ['The `__repr__()` and `__str__()` methods in Python are both used to return a string representation of an object. However, they have different purposes.\n\n* **`__repr__()`** is used to return a string representation that can be used to recreate the object. It is intended for developers, and should be unambiguous and informative.\n* **`__str__()`** is used to return a string representation that is readable by humans. It is intended for users, and can be more concise and user-friendly than the `__repr__()` representation.\n\nIf you do not define a `__str__()` method for a class, then the built-in object implementation will be used, which simply calls the `__repr__()` method.\n\nHere is an example:\n\n```python\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n```\n\nOutput:\n\n```\nA\nA str\n```\n\nAs you can see, the `__repr__()` representation is simply the string `"A"`, while the `__str__()` representation is more human-readable `"A str"`.\n\nIn general, it is a good practice to define both `__repr__()` and `__str__()` methods for your classes. The `__repr__()` method should be used whenever you need to get a string representation of an object for debugging or logging purposes. The `__str__()` method should be used whenever you need to get a string representation of an object for display to the user.'], [], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, 'A\nA str\n', '', 'what is the difference between __repr__ and __str__ in python?', None, 0, [], 1, '\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', None, None, None, None, None, None, None, 3, None, None, None, None, None, [None, None, True, None, None, None, None, None, None, None, None, None, 1, False, False], None, None, None, None, None, None, None, None, True]], None, None, None, None, None, None, None, None, None, None, None, None, None, [['\n\nclass A:\n    def __repr__(self):\n        return "A"\n\n    def __str__(self):\n        return "A str"\n\na = A()\nprint(repr(a))\nprint(str(a))\n', 0, 3, None, None, None, None, True]]], None, False, True, True, None, None, False]], ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599107262, 107302696], [599284034, 107655384]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1108.9832, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=Hm-dhkIkOCg36-UKGQBNL2fpqA7SbnEJn-zruThQA_tsUPeBUqssVQuSj2XXgDR3IUUUpaYpRVGpvhwTwADiY1WzJ3R1RiSJfb4oYzjtHqBF661XGf02ogkUmZEc0p1SshMone5vPIOuL8SDSBhsU0MwZKw7Yj2a_95cPvRWGE3DEfIw1JW2YNr-12qTXnsi'], None, None, 'NO', []], [None, ['c_903e0fc7657e7d7a', 'r_903e0fc7657e7dd7'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599107262, 107302696], [599284034, 107655384]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1108.9832, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=Hm-dhkIkOCg36-UKGQBNL2fpqA7SbnEJn-zruThQA_tsUPeBUqssVQuSj2XXgDR3IUUUpaYpRVGpvhwTwADiY1WzJ3R1RiSJfb4oYzjtHqBF661XGf02ogkUmZEc0p1SshMone5vPIOuL8SDSBhsU0MwZKw7Yj2a_95cPvRWGE3DEfIw1JW2YNr-12qTXnsi'], None, None, 'NO', None, ['Difference between __repr__ and __str__']]]
+
+# How many siblings does Jill Biden have? Please use citations.
+t8 = [[None, ['c_ec0ac24caebe37d5', 'r_ec0ac24caebe3bb6'], [['how many siblings does jill biden have?', 1], ['How much is Hunter Biden worth?', 4]], None, [['rc_ec0ac24caebe36dd', ['Jill Biden has four younger sisters:\r\n\r\n* Bonny Jean Biden\r\n* Barbara Biden\r\n* Valerie Biden Owens\r\n* Janine Biden\r\n\r\nSource: Jill Biden - Wikipedia: https://en.wikipedia.org/wiki/Jill_Biden\r\n\r\n> Jill Tracy Jacobs was born on June 3, 1951, in Hammonton, New Jersey. She is the oldest of five sisters. Her father, Donald Carl Jacobs, was a bank teller and U.S. Navy signalman during World War II who used the G.I. Bill to attend college and become an insurance salesman. Her mother, Bonny Jean Godfrey Jacobs, was a homemaker.'], [[[284, 425, ['http://en.wikipedia.org/wiki/Jill_Biden', '', '', None, '', None, None, '', False, '', '', '', '', '', '', ''], 1, 1, None, [1, 'normal_citation_datasets']], [196, 411, ['https://en.wikipedia.org/wiki/Jill_Biden#:~:text=Jill%20Tracy%20Jacobs%20was%20born,II%20who%20used%20the%20G.I.', '', '', '', '', None, None, '', False, '', '', '', '', '', '', ''], 2, 2, None, [1, 'normal_citation_datasets']]]], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], None, None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False], ['rc_ec0ac24caebe3abe', ["Jill Biden has **four** younger sisters. Their names are:\r\n\r\n* Jane Tracey\r\n* Nancy Craigen\r\n* Valarie Biden Owens\r\n* Bonny Jean Jacobs\r\n\r\nSource: Wikipedia: https://en.wikipedia.org/wiki/Jill_Biden\r\n\r\nPlease note that Jill Biden's sisters are not in the public eye, so there is relatively little information about them available online."], [], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], None, None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False], ['rc_ec0ac24caebe3e9f', ['Jill Biden has four younger sisters.\r\n\r\n* Bonny Jean Biden\r\n* Valarie Biden Owens\r\n* Jamie Biden\r\n* Jane Biden\r\n\r\nSource: Jill Biden on Wikipedia: https://en.wikipedia.org/wiki/Jill_Biden'], [], None, None, None, True, None, None, 'en', None, None, [None, None, None, None, None, [[None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, True, None, None, None, None, None, 1]]], None, None, None, None, None, None, None, None, None, None, None, None, None, [[]]], None, False, True, True, None, None, False]], ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599106274, 107300726], [599285022, 107657354]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1121.3759, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=cBEzFAyIXvfVsLp4YwuuP77lGp4L70x1pIeSKXiWU7OyfhAdSfr54ILOH3XJce3WfXIGOavCKI5WXfXA_76edgmTR7eGsKohTCmYYJm57zCNUhB0Twv9x6UxNLkPtZqvnqdsAMJoC5W-smDChKxUBNaMRAVUJJpP9EXt7cUHR92iM5HTKoJZoMVVoO63r63x'], None, None, 'NO', []], [None, ['c_ec0ac24caebe37d5', 'r_ec0ac24caebe3bb6'], None, None, None, ['Oslo, Norway', 'SWML_DESCRIPTION_FROM_YOUR_INTERNET_ADDRESS', False, [[[None, None, None, [[300, 200], None, None, [[599106274, 107300726], [599285022, 107657354]]]]], [[2, None, None, None, None, [None, None, None, [[[599195648, 107479040], 1121.3759, 1931574222, 1, 864863983]]]]], None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, None, [None, None, None, None, None, None, None, True]], '//www.google.com/maps/vt/data=cBEzFAyIXvfVsLp4YwuuP77lGp4L70x1pIeSKXiWU7OyfhAdSfr54ILOH3XJce3WfXIGOavCKI5WXfXA_76edgmTR7eGsKohTCmYYJm57zCNUhB0Twv9x6UxNLkPtZqvnqdsAMJoC5W-smDChKxUBNaMRAVUJJpP9EXt7cUHR92iM5HTKoJZoMVVoO63r63x'], None, None, 'NO', None, ["Jill Biden's 4 younger sisters"]]]
